requirement.txt
# ── Core simulation ────────────────────────────────────────────────────────────
mujoco>=3.1.0
gymnasium>=0.29.0

# ── RL framework ──────────────────────────────────────────────────────────────
stable-baselines3>=2.3.0
torch>=2.2.0

# ── Numerics / optimisation (MPC QP solver) ───────────────────────────────────
numpy>=1.26.0
scipy>=1.13.0

# ── Config ────────────────────────────────────────────────────────────────────
pyyaml>=6.0

# ── Logging / visualisation ───────────────────────────────────────────────────
tensorboard>=2.16.0

# ── Optional: faster QP solver for real-time MPC (replace scipy SLSQP) ───────
# osqp>=0.6.7
# qpsolvers[osqp]>=4.3.0
